+++ CHANGES.md
@@ -4,7 +4,8 @@
 
 - Purge use of remote access to IMDb data using Cinemagoer in favor of downloading the freely available non-commercial dataset directly from IMDb.
 - Use IMDb's Primary Title field when fetching the title.
-- Update bundled Qt 6 to version 6.11.
+- Update bundled Qt 6 to version 6.11 and Python to 3.14.
+- Use Hatch to build Windows executable using PyInstaller
 
 ## 1.0.0 (2023-11-22)